Added some impls and more tests

+++ src/main/java/com/oleksa/model/dao/RecordDao.java
@@ -6,9 +6,9 @@ import com.oleksa.model.entity.Record;
 
 public interface RecordDao extends CrudDao<Record, Long> {
 
-    List<Record> findAllByClientId(int clientId);
+    List<Record> findAllByClientId(long clientId);
 
-    List<Record> findAllByClientIdWithMaster(int clientId);
+    List<Record> findAllByClientIdWithMaster(long clientId);
 
 //    @Override
 //    Record create(Record t) throws RecordOccupiedException;
